move classes to their own files and update imports
